pyproject.toml
[project]
name = "agno-rag"
version = "0.1.0"
description = "Agno-native rebuild of the RAG-Anything ingestion/query pipeline"
requires-python = ">=3.11"
dependencies = [
  "agno",
  "openai",
  "python-dotenv",
  "pydantic>=2",
  "pydantic-settings>=2",
  "tiktoken",
  "pillow",
  "fastmcp",
  "lancedb",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]